requirements.txt
requests
numpy
pillow
selenium